chore: update composer dependencies (safe updates)
Updated 37 packages to latest patch/minor versions:
- brianium/paratest 7.12.0 → 7.16.1
- doctrine/orm 3.5.2 → 3.6.1
- phpstan/* extensions (5 packages)
- phpunit/phpunit 12.3.8 → 12.5.4
- rector/rector 2.1.6 → 2.3.1
- sentry/sentry-symfony 5.3.1 → 5.8.3
- symfony/* (various minor updates within 7.3.x)
- twig/twig 3.21.1 → 3.22.2

Fixed 8 sprintf type errors from stricter PHPStan rules:
- Cast floor() result to int in Entry::getDurationString()
- Cast DateTime::format('t') to int in EntryRepository
- Handle nullable getId() in OptimizedEntryRepository and JiraWorkLogService
- Cast float|int duration to int in performance test logging

+++ src/Entity/Entry.php
@@ -384,7 +384,7 @@ class Entry extends Base
     public function getDurationString(): string
     {
         $nMinutes = $this->getDuration();
-        $nHours = floor($nMinutes / 60);
+        $nHours = (int) floor($nMinutes / 60);
         $nMinutes %= 60;
 
         return sprintf('%02d:%02d', $nHours, $nMinutes);
